% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_geography.R
\docType{data}
\name{census_geography_list}
\alias{census_geography_list}
\title{census geography lookup}
\format{
An object of class \code{list} of length 5.
}
\usage{
census_geography_list
}
\description{
census geography lookup
}
\keyword{datasets}
