% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_geography.R
\name{census_to_cmhc_geocode}
\alias{census_to_cmhc_geocode}
\title{translate census to CMHC geocodes}
\usage{
census_to_cmhc_geocode(GeoUID)
}
\arguments{
\item{GeoUID}{Census GeoUID for CMA or CSD}
}
\description{
translate census to CMHC geocodes
}
