% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_geography.R
\docType{data}
\name{census_to_cmhc_translation}
\alias{census_to_cmhc_translation}
\title{census geography lookup}
\format{An object of class \code{list} of length 4.}
\usage{
census_to_cmhc_translation
}
\description{
census geography lookup
}
\keyword{datasets}
