% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{cmhc_csd_translation_data}
\alias{cmhc_csd_translation_data}
\title{A dataset with geo identifiers for CMHC and Census at the CSD level}
\description{
A dataset with geo identifiers for CMHC and Census at the CSD level
}
\author{
derived from CMHC geographic data
}
\keyword{data}
