% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{cmhc_csd_translation_data_2023}
\alias{cmhc_csd_translation_data_2023}
\title{A dataset with geographic identifiers for CMHC and Census at the CSD level for 2023 data portal version}
\format{
A tibble with Census CSD and CMHC METID geographic identifiers
}
\source{
Custom data extract from CMHC
}
\description{
A dataset with geographic identifiers for CMHC and Census at the CSD level for 2023 data portal version
}
\author{
derived from CMHC data
}
\keyword{data}
\keyword{internal}
