% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_tables.R
\docType{data}
\name{cmhc_data_table}
\alias{cmhc_data_table}
\title{Start to organize data better
CMCH identifier rules:
first number for survey
second number 1=fixed time, 2=time series
thrid number series
forth number (only for fixed time) geographic breakdown}
\format{
An object of class \code{list} of length 4.
}
\usage{
cmhc_data_table
}
\description{
Start to organize data better
CMCH identifier rules:
first number for survey
second number 1=fixed time, 2=time series
thrid number series
forth number (only for fixed time) geographic breakdown
}
\keyword{datasets}
