% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_geography.R
\name{cmhc_geo_level_for_census}
\alias{cmhc_geo_level_for_census}
\title{Get cmhc geo level for census geo_uid}
\usage{
cmhc_geo_level_for_census(GeoUID)
}
\arguments{
\item{GeoUID}{census region id}
}
\description{
Get cmhc geo level for census geo_uid
}
