% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc.R
\name{cmhc_geo_uid_for_ct}
\alias{cmhc_geo_uid_for_ct}
\title{Build ct_uid out of cmhc data}
\usage{
cmhc_geo_uid_for_ct(cma_header, GeoUID)
}
\arguments{
\item{cma_header}{Census CMA code without province part}

\item{GeoUID}{ct geo_uid}
}
\description{
Build ct_uid out of cmhc data
}
