% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_geography.R
\docType{data}
\name{cmhc_geography_list}
\alias{cmhc_geography_list}
\title{cmhc geography lookup for cma}
\format{
An object of class \code{list} of length 9.
}
\usage{
cmhc_geography_list
}
\description{
cmhc geography lookup for cma
}
\keyword{datasets}
