% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_geography.R
\docType{data}
\name{cmhc_geography_type_list}
\alias{cmhc_geography_type_list}
\title{cmhc geography type lookup}
\format{
An object of class \code{list} of length 6.
}
\usage{
cmhc_geography_type_list
}
\description{
cmhc geography type lookup
}
\keyword{datasets}
