% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc.R
\name{cmhc_primary_rental_params}
\alias{cmhc_primary_rental_params}
\title{Parameters for primary market rent change fixed sample data timeline}
\usage{
cmhc_primary_rental_params(geography_id = 2410, table_id = "2.2.12",
  default_data_field = "fixed_sample_rent_change_pct",
  data_source = "1")
}
\arguments{
\item{geography_id}{Geography for which to get the data}

\item{table_id}{cmhc table id}

\item{default_data_field}{data field}

\item{data_source}{not sure what this is}
}
\description{
Parameters for primary market rent change fixed sample data timeline
}
