% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cmhc_quality_labels}
\alias{cmhc_quality_labels}
\title{A list of CMHC quality indicators}
\format{
A named vector to translate internal CMHC quality indicators to plain text.
}
\usage{
cmhc_quality_labels
}
\description{
Data obtained via this package will automatically translate internal CMHC quality labels
using this translation vector, this named vector is useful when working with CMHC data obtained from other
sources like sporadic excel sheets or data scraped from PDF reports.
}
