% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_geography.R
\name{cmhc_region_params}
\alias{cmhc_region_params}
\title{cmhc geography params}
\usage{
cmhc_region_params(geography, type = "CMA")
}
\arguments{
\item{geography}{cmhc region id}

\item{type}{level of geography}
}
\description{
cmhc geography params
}
