% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_geography.R
\name{cmhc_region_params_from_census}
\alias{cmhc_region_params_from_census}
\title{census geography params}
\usage{
cmhc_region_params_from_census(GeoUID)
}
\arguments{
\item{GeoUID}{census region id}
}
\description{
census geography params
}
