% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc.R
\docType{data}
\name{cmhc_rms_geography_list}
\alias{cmhc_rms_geography_list}
\title{cmhc rms geography type lookup}
\format{An object of class \code{list} of length 4.}
\usage{
cmhc_rms_geography_list
}
\description{
cmhc rms geography type lookup
}
\keyword{datasets}
