% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc.R
\name{cmhc_snapshot_params}
\alias{cmhc_snapshot_params}
\title{Parameters for time series}
\usage{
cmhc_snapshot_params(
  table_id = "2.2.12",
  geography_id = 2410,
  geography_type = 3,
  breakdown_geography_type = "CSD",
  filter = list(),
  region = NA,
  year = 2017,
  month = 7,
  frequency = NA
)
}
\arguments{
\item{table_id}{CMHC table number to query}

\item{geography_id}{Geography for which to get the data}

\item{geography_type}{type corrsponding to geography}

\item{breakdown_geography_type}{breakdown geography type for results}

\item{filter}{optional filter for table}

\item{region}{can be used instead of the geography arguments, in particular in conjunction with cmhc_region_params_from_census}

\item{year}{year for which to query the data}

\item{month}{month for which to query the data}

\item{frequency}{frequency at which the data should be aggregated}
}
\description{
Parameters for time series
}
