% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_tables.R
\docType{data}
\name{cmhc_table_list}
\alias{cmhc_table_list}
\title{table lookup}
\format{An object of class \code{list} of length 44.}
\usage{
cmhc_table_list
}
\description{
table lookup
}
\keyword{datasets}
