% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc.R
\docType{data}
\name{cmhc_table_region_code}
\alias{cmhc_table_region_code}
\title{table region code lookup}
\format{An object of class \code{list} of length 5.}
\usage{
cmhc_table_region_code
}
\description{
table region code lookup
}
\keyword{datasets}
