% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_geography.R
\name{cmhc_to_census_geocode}
\alias{cmhc_to_census_geocode}
\title{translate CMHC to census geocodes}
\usage{
cmhc_to_census_geocode(GeoUID, CMA_GEOUID = NULL)
}
\arguments{
\item{GeoUID}{CMHC GeoUID for CMA or CSD or CT}

\item{CMA_GEOUID}{census geographic identifier for CMA}
}
\description{
translate CMHC to census geocodes
}
