% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc.R
\name{cmhc_vacancy_history_params}
\alias{cmhc_vacancy_history_params}
\title{Parameters for primary market vacancy data timeline}
\usage{
cmhc_vacancy_history_params(geography_id = 2410)
}
\arguments{
\item{geography_id}{Geography for which to get the data}
}
\description{
Parameters for primary market vacancy data timeline
}
