% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc.R
\name{cmhc_vacancy_params}
\alias{cmhc_vacancy_params}
\title{Parameters for primary market vacancy data buy survey zones}
\usage{
cmhc_vacancy_params()
}
\description{
Parameters for primary market vacancy data buy survey zones
}
