% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{date_from_cmhc_year_month}
\alias{date_from_cmhc_year_month}
\title{Convert CMHC month and date field to Date}
\usage{
date_from_cmhc_year_month(X1)
}
\arguments{
\item{X1}{month and date field as returned by CMHC}
}
\value{
a vector of dates
}
\description{
Convert CMHC month and date field to Date
}
