% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_geography.R
\name{download_geographies}
\alias{download_geographies}
\title{download local copy of cmhc geography data}
\usage{
download_geographies(base_directory = Sys.getenv("CMHC_CACHE_PATH"))
}
\arguments{
\item{base_directory}{local directory to hold cmhc geography data}
}
\description{
download local copy of cmhc geography data
}
