% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc.R
\docType{data}
\name{intended_market_rental}
\alias{intended_market_rental}
\title{intended market rental}
\format{An object of class \code{list} of length 1.}
\usage{
intended_market_rental
}
\description{
intended market rental
}
\keyword{datasets}
