% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_tables.R
\name{list_cmhc_breakdowns}
\alias{list_cmhc_breakdowns}
\title{List available CMHC breakdowns}
\usage{
list_cmhc_breakdowns(survey = NULL, series = NULL, dimension = NULL)
}
\arguments{
\item{survey}{Optional survey to filter by}

\item{series}{Optional series to filter by}

\item{dimension}{Optional dimension to filter by}
}
\value{
A data frame with survey names, series names, dimension names and series breakdowns.
}
\description{
List available CMHC breakdowns
}
\examples{
list_cmhc_breakdowns("Rms","Vacancy Rate","Bedroom Type")

}
