% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_tables.R
\name{list_cmhc_series}
\alias{list_cmhc_series}
\title{List available CMHC series}
\usage{
list_cmhc_series(survey = NULL)
}
\arguments{
\item{survey}{Optional survey to filter by}
}
\value{
A data frame with survey names, and available series names.
}
\description{
List available CMHC series
}
\examples{
list_cmhc_series("Rms")

}
