% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_tables.R
\name{list_cmhc_surveys}
\alias{list_cmhc_surveys}
\title{List available CMHC surveys}
\usage{
list_cmhc_surveys()
}
\value{
A data frame with available survey names.
}
\description{
List available CMHC surveys
}
\examples{
list_cmhc_surveys()

}
