% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_tables.R
\name{list_cmhc_tables}
\alias{list_cmhc_tables}
\title{List available CMHC tables}
\usage{
list_cmhc_tables(short = TRUE)
}
\arguments{
\item{short}{Logical, determines how much detail is returned. Default is `TRUE`.}
}
\value{
A tibble listing all available CMHC data tables
}
\description{
List available CMHC tables
}
\examples{
list_cmhc_tables()

}
