% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_tables.R
\name{select_cmhc_table}
\alias{select_cmhc_table}
\title{Interactive table selector}
\usage{
select_cmhc_table()
}
\value{
A string containing the function call to access the selected table
}
\description{
Interactive table selector
}
\examples{
\dontrun{
select_cmhc_table()
}
}
