% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_settings.R
\name{set_cache_path}
\alias{set_cache_path}
\title{Set persistent cmhc cache location}
\usage{
set_cache_path(cache_path, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{cache_path}{a local directory to use for saving cached data}

\item{overwrite}{Option to overwrite any existing cache path already stored locally.}

\item{install}{Option to install permanently for use across sessions.}
}
\description{
cmhc provides access to custom cmhc geographies, these are large files and should be stored in a
permanent location.
}
\examples{
\dontrun{
# This sets the cache path for the duration of the current session
set_cache_path("~/cmhc_cache")

# This will set the cache path permanently until overwritten again
set_cache_path("~/cmhc_cache", install = TRUE)
}

}
