% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_settings.R
\name{show_cmhc_cache_path}
\alias{show_cmhc_cache_path}
\title{View saved cache directory path}
\usage{
show_cmhc_cache_path()
}
\value{
a character string with the CMHC cache path
}
\description{
View saved cache path
}
\examples{
show_cmhc_cache_path()

}
