% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc.R
\name{standardize_cmhc_data}
\alias{standardize_cmhc_data}
\title{convert CMHC data to long form with quality indicator}
\usage{
standardize_cmhc_data(data)
}
\arguments{
\item{data}{cmhc data frame as returned by get_cmhc}
}
\value{
a long form data frame with values in Value column and quality indicators in Quality column
}
\description{
convert CMHC data to long form with quality indicator
}
\examples{
\dontrun{
query_params <- cmhc_timeseries_params(cmhc_table_list$`Rms Rent Change Time Series`,
                                       region=cmhc_region_params_from_census("59933"))
data <- get_cmhc(query_params)

d <- standardize_cmhc_data(data)

}
}
