% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{cmhc_csd_translation_data}
\alias{cmhc_csd_translation_data}
\title{A dataset with geographic identifiers for CMHC and Census at the CSD level}
\format{
A tibble with Census and CMHC geographic identifiers
}
\source{
Custom data extract from CMHC
}
\description{
A dataset with geographic identifiers for CMHC and Census at the CSD level
}
\author{
derived from CMHC geographic data
}
\keyword{data}
