% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc.R
\docType{data}
\name{cmhc_geography_csd_list}
\alias{cmhc_geography_csd_list}
\title{census geography cities}
\format{An object of class \code{list} of length 15.}
\usage{
cmhc_geography_csd_list
}
\description{
census geography cities
}
\keyword{datasets}
