% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc.R
\docType{data}
\name{cmhc_geography_list}
\alias{cmhc_geography_list}
\title{geography lookup}
\format{An object of class \code{list} of length 3.}
\usage{
cmhc_geography_list
}
\description{
geography lookup
}
\keyword{datasets}
