% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{cmhc_quality_labels}
\alias{cmhc_quality_labels}
\title{CMHC quality indicators}
\format{
An object of class \code{character} of length 4.
}
\usage{
cmhc_quality_labels
}
\description{
CMHC quality indicators
}
\keyword{datasets}
