% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc.R
\name{cmhc_rent_change_history_params}
\alias{cmhc_rent_change_history_params}
\title{Parameters for primary market rent change fixed sample data timeline}
\usage{
cmhc_rent_change_history_params(geography_id = 2410)
}
\description{
Parameters for primary market rent change fixed sample data timeline
}
