% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc.R
\name{cmhc_snapshot_params}
\alias{cmhc_snapshot_params}
\title{Parameters for time series}
\usage{
cmhc_snapshot_params(table_id = "2.2.12", geography_id = 2410,
  geography_type = 3, breakdown_geography_type = "CSD", filter = list(),
  region = NA, year = 2017, month = 7, frequency = NA)
}
\arguments{
\item{table_id}{CMHC table id}

\item{geography_id}{Geography for which to get the data}

\item{geography_type}{type corrsponding to geography}
}
\description{
Parameters for time series
}
