% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_geography.R
\name{cmhc_to_census_geocode}
\alias{cmhc_to_census_geocode}
\title{translate CMHC to census geocodes}
\usage{
cmhc_to_census_geocode(GeoUID, parent_region = NULL)
}
\arguments{
\item{GeoUID}{CMHC GeoUID for CMA or CSD or CT}

\item{parent_region}{census geographic identifier to identify CMA if needed}
}
\description{
translate CMHC to census geocodes
}
