% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc.R
\name{get_cmhc}
\alias{get_cmhc}
\title{Hack into CMHC data portal to pull out csv tables}
\usage{
get_cmhc(query_params)
}
\arguments{
\item{query_params}{Query parameters to be sent to CMHC.

returns dataframe with CMHC data, tile and subtitle are set as attributes}
}
\description{
Hack into CMHC data portal to pull out csv tables
}
\examples{
dat=get_cmhc(cmhc_rent_change_history_params())
}
\keyword{api}
\keyword{canada}
\keyword{cmhc}
\keyword{data}
\keyword{hack}
