% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc.R
\name{get_cmhc}
\alias{get_cmhc}
\title{Hack into CMHC data portal to pull out csv tables}
\usage{
get_cmhc(query_params)
}
\arguments{
\item{query_params}{Query parameters to be sent to CMHC.

returns dataframe with CMHC data, tile and subtitle are set as attributes}
}
\description{
Hack into CMHC data portal to pull out csv tables
}
\examples{
table <- cmhc_table_list$`Rms Vacancy Rate Time Series`
region - cmhc_region_params_from_census("59933")
dat <- get_cmhc(cmhc_timeseries_params(table_id = table, region = region))

}
\keyword{api}
\keyword{canada}
\keyword{cmhc}
\keyword{data}
\keyword{hack}
