% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmhc_geography.R
\name{get_cmhc_geography}
\alias{get_cmhc_geography}
\title{get cmhc geographies, requires data has been downloaded with `download_geographies()` first}
\usage{
get_cmhc_geography(level = c("CT", "ZONE", "NBHD", "CSD", "MET"),
  base_directory = getOption("cmhc.cache_path"))
}
\arguments{
\item{level}{aggreagtion level for geographic data, one of "CT","ZONE","NBHD","CSD","MET"}

\item{base_directory}{local directory to hold cmhc geography data}
}
\description{
get cmhc geographies, requires data has been downloaded with `download_geographies()` first
}
